Fix all moved component relative file paths
